Fixes the problem of incomplete release titles.
Example:
Mr.Peabody.And.Sherman.2014.NTSC.DVDR-JFKDVD results in the title "Mr Peabody And"

What happens:
"DVD" is recognized by the regular expression and is removed from the title, but the "R" is kept in the title. This leads to:
Mr.Peabody.And.Sherman.2014.NTSC.DVDR-JFKDVD
Mr.Peabody.And.Sherman.2014.NTSCR-JFKDVD
...
Mr.Peabody.And.ShermanR

Matching the parts of the title only the first 3 words are recognized.

Two issues lead to this behavior:
- Minor Issue: DVDR source is missing
- Major: Not the full parts between Dots are considered, but also substrings because the final part of the regex is optional "?"

Fix:
Make the last part of the regex not optional but check for ".", "-" oder END-OF-STRING ($).